test(ci): enforce platform-owned test execution
