+++ DownKyi.Core/Properties/AssemblyInfo.cs
@@ -1,6 +1,7 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("DownKyi.Core.Tests")]
+[assembly: InternalsVisibleTo("DownKyi.Windows.Tests")]
 [assembly: InternalsVisibleTo("DownKyi.Tests")]
 [assembly: InternalsVisibleTo("DownKyi.BenchmarkCases")]
 [assembly: InternalsVisibleTo("DownKyi.SystemBenchmarks")]
