+++ DownKyi.sln
@@ -58,6 +58,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tools", "tools", "{07C2787E
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DownKyi.AssemblyLifecycleProbe", "tools\DownKyi.AssemblyLifecycleProbe\DownKyi.AssemblyLifecycleProbe.csproj", "{14208581-8B0E-4E8F-988A-19ADB2EDFCE8}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DownKyi.MacOS.Tests", "tests\DownKyi.MacOS.Tests\DownKyi.MacOS.Tests.csproj", "{55B2C90F-9EF8-40BE-9156-EF7384B2601B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DownKyi.Windows.Tests", "tests\DownKyi.Windows.Tests\DownKyi.Windows.Tests.csproj", "{D1F36CA9-2EF9-433F-8AA6-0D89077B2458}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -284,6 +288,30 @@ Global
 		{14208581-8B0E-4E8F-988A-19ADB2EDFCE8}.Release|x64.Build.0 = Release|Any CPU
 		{14208581-8B0E-4E8F-988A-19ADB2EDFCE8}.Release|x86.ActiveCfg = Release|Any CPU
 		{14208581-8B0E-4E8F-988A-19ADB2EDFCE8}.Release|x86.Build.0 = Release|Any CPU
+		{55B2C90F-9EF8-40BE-9156-EF7384B2601B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{55B2C90F-9EF8-40BE-9156-EF7384B2601B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{55B2C90F-9EF8-40BE-9156-EF7384B2601B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{55B2C90F-9EF8-40BE-9156-EF7384B2601B}.Debug|x64.Build.0 = Debug|Any CPU
+		{55B2C90F-9EF8-40BE-9156-EF7384B2601B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{55B2C90F-9EF8-40BE-9156-EF7384B2601B}.Debug|x86.Build.0 = Debug|Any CPU
+		{55B2C90F-9EF8-40BE-9156-EF7384B2601B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{55B2C90F-9EF8-40BE-9156-EF7384B2601B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{55B2C90F-9EF8-40BE-9156-EF7384B2601B}.Release|x64.ActiveCfg = Release|Any CPU
+		{55B2C90F-9EF8-40BE-9156-EF7384B2601B}.Release|x64.Build.0 = Release|Any CPU
+		{55B2C90F-9EF8-40BE-9156-EF7384B2601B}.Release|x86.ActiveCfg = Release|Any CPU
+		{55B2C90F-9EF8-40BE-9156-EF7384B2601B}.Release|x86.Build.0 = Release|Any CPU
+		{D1F36CA9-2EF9-433F-8AA6-0D89077B2458}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D1F36CA9-2EF9-433F-8AA6-0D89077B2458}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D1F36CA9-2EF9-433F-8AA6-0D89077B2458}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D1F36CA9-2EF9-433F-8AA6-0D89077B2458}.Debug|x64.Build.0 = Debug|Any CPU
+		{D1F36CA9-2EF9-433F-8AA6-0D89077B2458}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D1F36CA9-2EF9-433F-8AA6-0D89077B2458}.Debug|x86.Build.0 = Debug|Any CPU
+		{D1F36CA9-2EF9-433F-8AA6-0D89077B2458}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D1F36CA9-2EF9-433F-8AA6-0D89077B2458}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D1F36CA9-2EF9-433F-8AA6-0D89077B2458}.Release|x64.ActiveCfg = Release|Any CPU
+		{D1F36CA9-2EF9-433F-8AA6-0D89077B2458}.Release|x64.Build.0 = Release|Any CPU
+		{D1F36CA9-2EF9-433F-8AA6-0D89077B2458}.Release|x86.ActiveCfg = Release|Any CPU
+		{D1F36CA9-2EF9-433F-8AA6-0D89077B2458}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -305,5 +333,7 @@ Global
 		{8340BC87-EAA0-4C15-9795-0A33B8139A05} = {66320409-64EC-F7C5-3DEF-65E7510DAAD1}
 		{34AA2BE0-A730-4CC1-8250-0B524B6EEB67} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
 		{14208581-8B0E-4E8F-988A-19ADB2EDFCE8} = {07C2787E-EAC7-C090-1BA3-A61EC2A24D84}
+		{55B2C90F-9EF8-40BE-9156-EF7384B2601B} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
+		{D1F36CA9-2EF9-433F-8AA6-0D89077B2458} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
 	EndGlobalSection
 EndGlobal
