+++ tests/DownKyi.Core.Tests/DownKyi.Core.Tests.csproj
@@ -7,6 +7,7 @@
     <Nullable>enable</Nullable>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <IsPackable>false</IsPackable>
+    <DownKyiTestPlatform>cross-platform</DownKyiTestPlatform>
   </PropertyGroup>
 
   <ItemGroup>
