+++ tests/DownKyi.MacOS.Tests/DownKyi.MacOS.Tests.csproj
@@ -1,27 +1,23 @@
 <Project Sdk="Microsoft.NET.Sdk">
+
   <PropertyGroup>
     <TargetFramework>net10.0</TargetFramework>
     <OutputType>Exe</OutputType>
     <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
     <IsPackable>false</IsPackable>
-    <DownKyiTestPlatform>cross-platform</DownKyiTestPlatform>
+    <DownKyiTestPlatform>macos</DownKyiTestPlatform>
   </PropertyGroup>
+
   <ItemGroup>
     <PackageReference Include="coverlet.collector" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" />
-    <PackageReference Include="Microsoft.Extensions.Http" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" />
     <PackageReference Include="xunit.v3" />
     <PackageReference Include="xunit.runner.visualstudio" />
   </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\DownKyi.Application\DownKyi.Application.csproj" />
-    <ProjectReference Include="..\..\src\DownKyi.Infrastructure\DownKyi.Infrastructure.csproj" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="Fixtures\*.db" CopyToOutputDirectory="PreserveNewest" />
-  </ItemGroup>
+
   <ItemGroup>
     <Using Include="Xunit" />
   </ItemGroup>
+
 </Project>
