+++ tests/DownKyi.Tests/DownKyi.Tests.csproj
@@ -6,6 +6,7 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <IsPackable>false</IsPackable>
+    <DownKyiTestPlatform>cross-platform</DownKyiTestPlatform>
   </PropertyGroup>
 
   <ItemGroup>
